requirements.txt
markdown==2.6.6
python-slugify==1.2.0
